# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/reseller/clients:
    post:
      operationId: createResellerClient
      tags: [admin]
      summary: Create a new VPN client via RoyaltyKey API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResellerClientBody"
      responses:
        "201":
          description: Created VPN client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResellerClient"
  /admin/subscriptions/{id}/renew:
    post:
      operationId: renewSubscription
      tags: [admin]
      summary: Renew a subscription via RoyaltyKey API
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewSubscriptionBody"
      responses:
        "200":
          description: Renewed subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResellerClient"
  /admin/subscriptions/{id}:
    delete:
      operationId: deleteSubscription
      tags: [admin]
      summary: Deactivate and delete a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
components:
  schemas:
...
    ResellerClient:
      type: object
      required: [id, vpnKey, expiresAt, tariffId]
      properties:
        id:
          type: string
        vpnKey:
          type: string
        expiresAt:
          type: string
        tariffId:
          type: string
    CreateResellerClientBody:
      type: object
      required: [tariffId]
      properties:
        tariffId:
          type: string
        label:
          type: string
          description: Optional label/name for the client
    RenewSubscriptionBody:
      type: object
      required: [tariffId]
      properties:
        tariffId:
          type: string
...