# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get dashboard statistics
 */
export const GetAdminStatsResponse = zod.object({
  "totalUsers": zod.number(),
  "totalRevenue": zod.number(),
  "activeSubscriptions": zod.number(),
  "expiredSubscriptions": zod.number(),
  "trialUsers": zod.number(),
  "pendingPayments": zod.number(),
  "paidPayments": zod.number(),
  "revenueToday": zod.number(),
  "newUsersToday": zod.number(),
  "recentPayments": zod.array(zod.object({
  "id": zod.string(),
  "telegramUserId": zod.string(),
  "username": zod.string().nullish(),
  "tariffId": zod.string(),
  "amount": zod.number(),
  "status": zod.string()
})),
  "activeToday": zod.number(),
  "activeWeek": zod.number(),
  "activeMonth": zod.number()
})


/**
 * @summary List all users
 */
export const listAdminUsersQueryPageDefault = 1;
export const listAdminUsersQueryLimitDefault = 20;

export const ListAdminUsersQueryParams = zod.object({
  "page": zod.coerce.number().default(listAdminUsersQueryPageDefault),
  "limit": zod.coerce.number().default(listAdminUsersQueryLimitDefault),
  "search": zod.coerce.string().optional()
})

export const ListAdminUsersResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.string(),
  "username": zod.string().nullish(),
  "balance": zod.number(),
  "hasUsedTrial": zod.boolean(),
  "referredById": zod.string().nullish(),
  "subscriptionCount": zod.number(),
  "lastActivityAt": zod.string().nullish()
})),
  "total": zod.number(),
  "page": zod.number(),
  "limit": zod.number()
})


/**
 * @summary List all payments
 */
export const listAdminPaymentsQueryPageDefault = 1;
export const listAdminPaymentsQueryLimitDefault = 20;

export const ListAdminPaymentsQueryParams = zod.object({
  "page": zod.coerce.number().default(listAdminPaymentsQueryPageDefault),
  "limit": zod.coerce.number().default(listAdminPaymentsQueryLimitDefault),
  "status": zod.enum(['pending', 'paid', 'failed']).optional()
})

export const ListAdminPaymentsResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.string(),
  "telegramUserId": zod.string(),
  "username": zod.string().nullish(),
  "tariffId": zod.string(),
  "amount": zod.number(),
  "status": zod.string()
})),
  "total": zod.number(),
  "page": zod.number(),
  "limit": zod.number()
})


/**
 * @summary List all subscriptions
 */
export const listAdminSubscriptionsQueryPageDefault = 1;
export const listAdminSubscriptionsQueryLimitDefault = 20;

export const ListAdminSubscriptionsQueryParams = zod.object({
  "page": zod.coerce.number().default(listAdminSubscriptionsQueryPageDefault),
  "limit": zod.coerce.number().default(listAdminSubscriptionsQueryLimitDefault),
  "active": zod.coerce.boolean().optional()
})

export const ListAdminSubscriptionsResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.string(),
  "telegramUserId": zod.string(),
  "username": zod.string().nullish(),
  "vpnKey": zod.string().optional(),
  "tariffId": zod.string(),
  "expiresAt": zod.string(),
  "isActive": zod.boolean()
})),
  "total": zod.number(),
  "page": zod.number(),
  "limit": zod.number()
})


/**
 * @summary List all promocodes
 */
export const ListAdminPromocodesResponseItem = zod.object({
  "id": zod.string(),
  "bonusAmount": zod.number(),
  "maxUses": zod.number(),
  "usesCount": zod.number()
})
export const ListAdminPromocodesResponse = zod.array(ListAdminPromocodesResponseItem)


/**
 * @summary Create a new promocode
 */
export const CreateAdminPromocodeBody = zod.object({
  "id": zod.string(),
  "bonusAmount": zod.number(),
  "maxUses": zod.number()
})

export const CreateAdminPromocodeResponse = zod.object({
  "id": zod.string(),
  "bonusAmount": zod.number(),
  "maxUses": zod.number(),
  "usesCount": zod.number()
})


/**
 * @summary Get public auth config (bot username)
 */
export const GetAuthConfigResponse = zod.object({
  "botUsername": zod.string()
})


/**
 * @summary Login via Telegram auth data (Login Widget)
 */
export const TelegramLoginBody = zod.object({
  "id": zod.number(),
  "hash": zod.string(),
  "auth_date": zod.number(),
  "first_name": zod.string().optional(),
  "last_name": zod.string().optional(),
  "username": zod.string().optional(),
  "photo_url": zod.string().optional()
})

export const TelegramLoginResponse = zod.object({
  "id": zod.number(),
  "firstName": zod.string(),
  "username": zod.string().optional(),
  "lastName": zod.string().optional()
})


/**
 * @summary Get current authenticated user
 */
export const GetAuthMeResponse = zod.object({
  "id": zod.number(),
  "firstName": zod.string(),
  "username": zod.string().optional(),
  "lastName": zod.string().optional()
})


/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  "ok": zod.boolean()
})


/**
 * @summary Get RoyaltyKey reseller balance and discount
 */
export const GetResellerProfileResponse = zod.object({
  "balance": zod.number().describe('Reseller balance in rubles'),
  "discount": zod.number().describe('Current wholesale discount percentage (0-40)')
})


/**
 * @summary Create a new VPN client via RoyaltyKey API
 */
export const CreateResellerClientBody = zod.object({
  "tariffId": zod.string(),
  "label": zod.string().optional().describe('Optional label\/name for the client')
})

export const CreateResellerClientResponse = zod.object({
  "id": zod.string(),
  "vpnKey": zod.string(),
  "expiresAt": zod.string(),
  "tariffId": zod.string()
})


/**
 * @summary Renew a subscription via RoyaltyKey API
 */
export const RenewSubscriptionParams = zod.object({
  "id": zod.coerce.string()
})

export const RenewSubscriptionBody = zod.object({
  "tariffId": zod.string()
})

export const RenewSubscriptionResponse = zod.object({
  "id": zod.string(),
  "vpnKey": zod.string(),
  "expiresAt": zod.string(),
  "tariffId": zod.string()
})


/**
 * @summary Deactivate and delete a subscription
 */
export const DeleteSubscriptionParams = zod.object({
  "id": zod.coerce.string()
})

export const DeleteSubscriptionResponse = zod.object({
  "ok": zod.boolean()
})


/**
 * @summary Adjust user balance
 */
export const AdjustUserBalanceParams = zod.object({
  "userId": zod.coerce.string()
})

export const AdjustUserBalanceBody = zod.object({
  "delta": zod.number()
})

export const AdjustUserBalanceResponse = zod.object({
  "id": zod.string(),
  "username": zod.string().nullish(),
  "balance": zod.number(),
  "hasUsedTrial": zod.boolean(),
  "referredById": zod.string().nullish(),
  "subscriptionCount": zod.number()
})